requirements.txt
flask
openai
pytube
python-dotenv
flask-cors
yt-dlp
requests
gunicorn